requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# NVIDIA HPC and GPU Computing
nvidia-ml-py3==7.352.0
cupy-cuda12x==12.2.0
numba==0.58.1

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
datasets==2.15.0
accelerate==0.25.0

# Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10

# OCR and Layout Analysis
layoutlmft==1.0.0
donut-python==1.0.8
paddlepaddle-gpu==2.5.2
paddleocr==2.7.0

# NLP and Language Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9

# Spanish Language Support
https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.7.0/es_core_news_lg-3.7.0.tar.gz#egg=es_core_news_lg

# Vector Database and Embeddings
weaviate-client==3.25.3
sentence-transformers==2.2.2
faiss-gpu==1.7.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1

# Stream Processing
kafka-python==2.0.2
confluent-kafka==2.3.0

# API and Web
requests==2.31.0
httpx==0.25.2
websockets==12.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Docker and Deployment
docker==6.1.3
kubernetes==28.1.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1 